my_base2/QFLBackdoorAttacks/requirements.txt
# Core requirements for QFLBackdoorAttacks
numpy>=1.21
scipy>=1.7
pandas>=1.3
matplotlib>=3.4
seaborn>=0.11
torch>=1.10
torchvision>=0.11
scikit-learn>=0.24
PyYAML>=5.4
loguru>=0.5
# Quantum
pennylane>=0.25
qiskit>=0.36
qiskit-aer>=0.10
# For IBMQ provider (optional)
qiskit-ibmq-provider>=0.19
# For plotting and figures
tqdm>=4.62
# For ablation and sweeps
optuna>=2.10
# For reproducibility
python-dotenv>=0.19
# For Colab compatibility
ipykernel
